+++ controllers/ConvertController.php
@@ -7,19 +7,24 @@ use yii\web\HttpException;
 use yii\helpers\Url;
 use humhub\modules\file\models\File;
 use humhub\modules\file\libs\FileHelper;
-use humhub\modules\onlyoffice\Module;
 use humhub\modules\onlyoffice\components\BaseFileController;
+use \humhub\components\Module;
 
 class ConvertController extends BaseFileController
 {
+
+    /**
+     * @var Module
+     */
+    public $module;
 
     public function init()
     {
         parent::init();
 
-        $module = Yii::$app->getModule('onlyoffice');
+        $this->module = Yii::$app->getModule('onlyoffice');
 
-        if (!$module->canConvert($this->file)) {
+        if (!$this->module->canConvert($this->file)) {
             throw new HttpException('400', 'Could not convert this file');
         }
     }
@@ -32,9 +37,8 @@ class ConvertController extends BaseFileController
     public function actionConvert($guid, $ts, $newName)
     {
         Yii::$app->response->format = 'json';
-        $module = Yii::$app->getModule('onlyoffice');
 
-        $json = $module->convertService($this->file, $ts);
+        $json = $this->module->convertService($this->file, $ts);
 
         if (!empty($json["endConvert"]) && $json["endConvert"]) {
             $this->saveFileReplace($json["fileUrl"], $newName);
@@ -44,7 +48,7 @@ class ConvertController extends BaseFileController
     }
 
     private function saveFileReplace($url, $newName) {
-        $content = file_get_contents($url);
+        $content = $this->module->request($url)->getBody();
 
         $this->file->store->setContent($content);
         $this->file->updateAttributes([