+++ models/ConfigureForm.php
@@ -12,6 +12,7 @@ class ConfigureForm extends \yii\base\Model
 {
 
     public $serverUrl;
+    public $verifyPeerOff;
     public $jwtSecret;
     public $internalServerUrl;
     public $storageUrl;
@@ -23,6 +24,7 @@ class ConfigureForm extends \yii\base\Model
     {
         return [
             ['serverUrl', 'string'],
+            ['verifyPeerOff', 'boolean'],
             ['jwtSecret', 'string'],
             ['internalServerUrl', 'string'],
             ['storageUrl', 'string'],
@@ -36,6 +38,7 @@ class ConfigureForm extends \yii\base\Model
     {
         return [
             'serverUrl' => Yii::t('OnlyofficeModule.base', 'Hostname'),
+            'verifyPeerOff' => Yii::t('OnlyofficeModule.base', 'Disable certificate verification (insecure)'),
             'jwtSecret' => Yii::t('OnlyofficeModule.base', 'JWT Secret'),
             'internalServerUrl' => Yii::t('OnlyofficeModule.base', 'ONLYOFFICE Docs address for internal requests from the server'),
             'storageUrl' => Yii::t('OnlyofficeModule.base', 'Server address for internal requests from ONLYOFFICE Docs'),
@@ -58,6 +61,7 @@ class ConfigureForm extends \yii\base\Model
     public function loadSettings()
     {
         $this->serverUrl = Yii::$app->getModule('onlyoffice')->settings->get('serverUrl');
+        $this->verifyPeerOff = (boolean)Yii::$app->getModule('onlyoffice')->settings->get('verifyPeerOff');
         $this->jwtSecret = Yii::$app->getModule('onlyoffice')->settings->get('jwtSecret');
         $this->internalServerUrl = Yii::$app->getModule('onlyoffice')->settings->get('internalServerUrl');
 
@@ -67,6 +71,7 @@ class ConfigureForm extends \yii\base\Model
     public function save()
     {
         Yii::$app->getModule('onlyoffice')->settings->set('serverUrl', rtrim($this->serverUrl, '/'));
+        Yii::$app->getModule('onlyoffice')->settings->set('verifyPeerOff', $this->verifyPeerOff);
         Yii::$app->getModule('onlyoffice')->settings->set('jwtSecret', $this->jwtSecret);
         Yii::$app->getModule('onlyoffice')->settings->set('internalServerUrl', rtrim($this->internalServerUrl, '/'));
         Yii::$app->getModule('onlyoffice')->settings->set('storageUrl', rtrim($this->storageUrl, '/'));